# openapi/payment-sim.yaml
openapi: 3.1.0
info:
  title: Payment Simulator API
  description: Traffic Tacos 결제 모의 서비스 API
  version: 1.0.0
  contact:
    name: Traffic Tacos Team

servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.traffic-tacos.com
    description: Production server

security:
  - IdempotencyKey: []

paths:
  /v1/sim/intent:
    post:
      summary: 결제 인텐트 생성
      description: 새로운 결제 인텐트를 생성하고 웹훅을 예약합니다.
      operationId: createPaymentIntent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentIntentRequest'
      responses:
        '200':
          description: 결제 인텐트 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePaymentIntentResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 멱등성 키 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 내부 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/sim/webhook/test:
    post:
      summary: 웹훅 테스트 발송
      description: 지정된 URL로 테스트 웹훅을 수동으로 발송합니다.
      operationId: testWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestWebhookRequest'
      responses:
        '202':
          description: 웹훅 발송 요청 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestWebhookResponse'
        '404':
          description: 결제 인텐트 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/sim/intents/{paymentIntentId}:
    get:
      summary: 결제 인텐트 조회
      description: 결제 인텐트의 현재 상태를 조회합니다.
      operationId: getPaymentIntent
      parameters:
        - name: paymentIntentId
          in: path
          required: true
          schema:
            type: string
          description: 결제 인텐트 ID
      responses:
        '200':
          description: 결제 인텐트 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntent'
        '404':
          description: 결제 인텐트 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /healthz:
    get:
      summary: 헬스 체크
      description: 서비스 상태를 확인합니다.
      operationId: healthCheck
      responses:
        '200':
          description: 서비스 정상
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /readyz:
    get:
      summary: 레디니스 체크
      description: 서비스 준비 상태를 확인합니다.
      operationId: readinessCheck
      responses:
        '200':
          description: 서비스 준비 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /metrics:
    get:
      summary: 메트릭스
      description: Prometheus 메트릭스를 반환합니다.
      operationId: metrics
      responses:
        '200':
          description: Prometheus 메트릭스
          content:
            text/plain:
              schema:
                type: string

components:
  securitySchemes:
    IdempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key
      description: 요청의 멱등성을 보장하기 위한 키

  schemas:
    CreatePaymentIntentRequest:
      type: object
      required:
        - reservation_id
        - amount
        - scenario
        - webhook_url
      properties:
        reservation_id:
          type: string
          description: 예약 ID
          example: "rsv_abc123"
        amount:
          type: integer
          description: 결제 금액 (원)
          minimum: 1
          example: 120000
        currency:
          type: string
          description: 통화 코드
          default: "KRW"
          example: "KRW"
        scenario:
          type: string
          enum: [approve, fail, delay, random]
          description: 결제 시뮬레이션 시나리오
          example: "approve"
        delay_ms:
          type: integer
          description: 지연 시간 (밀리초)
          minimum: 0
          example: 0
        webhook_url:
          type: string
          format: uri
          description: 웹훅 콜백 URL
          example: "https://reservation-api.example.com/internal/payment/webhook"
        metadata:
          type: object
          description: 추가 메타데이터
          additionalProperties: true
          example:
            user_id: "u123"
            event_id: "evt_2025_1001"

    CreatePaymentIntentResponse:
      type: object
      required:
        - payment_intent_id
        - status
        - next
      properties:
        payment_intent_id:
          type: string
          description: 생성된 결제 인텐트 ID
          example: "pay_xyz789"
        status:
          type: string
          enum: [PENDING]
          description: 결제 인텐트 상태
          example: "PENDING"
        next:
          type: string
          enum: [webhook]
          description: 다음 단계
          example: "webhook"

    TestWebhookRequest:
      type: object
      required:
        - payment_intent_id
        - type
        - webhook_url
      properties:
        payment_intent_id:
          type: string
          description: 결제 인텐트 ID
          example: "pay_xyz789"
        type:
          type: string
          enum: [payment.approved, payment.failed]
          description: 웹훅 타입
          example: "payment.approved"
        webhook_url:
          type: string
          format: uri
          description: 웹훅 발송 URL
          example: "https://test.example.com/webhook"

    TestWebhookResponse:
      type: object
      required:
        - sent
      properties:
        sent:
          type: boolean
          description: 웹훅 발송 여부
          example: true

    PaymentIntent:
      type: object
      properties:
        payment_intent_id:
          type: string
          description: 결제 인텐트 ID
          example: "pay_xyz789"
        reservation_id:
          type: string
          description: 예약 ID
          example: "rsv_abc123"
        amount:
          type: integer
          description: 결제 금액
          example: 120000
        currency:
          type: string
          description: 통화 코드
          example: "KRW"
        scenario:
          type: string
          description: 시뮬레이션 시나리오
          example: "approve"
        delay_ms:
          type: integer
          description: 지연 시간
          example: 0
        webhook_url:
          type: string
          description: 웹훅 URL
          example: "https://reservation-api.example.com/internal/payment/webhook"
        metadata:
          type: object
          description: 메타데이터
          additionalProperties: true
        status:
          type: string
          enum: [PENDING, APPROVED, FAILED]
          description: 결제 상태
          example: "APPROVED"
        created_at:
          type: string
          format: date-time
          description: 생성 시간
          example: "2024-01-01T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: 수정 시간
          example: "2024-01-01T12:00:00Z"
        last_attempt_at:
          type: string
          format: date-time
          description: 마지막 시도 시간
          nullable: true
        attempt_count:
          type: integer
          description: 시도 횟수
          minimum: 0
          example: 1

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: 에러 코드
              example: "VALIDATION_FAILED"
            message:
              type: string
              description: 에러 메시지
              example: "Invalid JSON payload"
            trace_id:
              type: string
              description: 트레이스 ID
              nullable: true
              example: "abc123def456"

    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [ok, ready]
          description: 서비스 상태
          example: "ok"